refactor: UserResponse -> Detail,List Response로 쪼개기
